Интерактор и репозиторий Favorite переименованы и преобразованы под задачи FavoriteFragment. Убраны ненужные suspend.
