+++ app/src/main/java/ru/practicum/android/diploma/domain/favorite/interactors/GetFavoriteVacanciesInteractorListImpl.kt
@@ -1,0 +1,13 @@
+package ru.practicum.android.diploma.domain.favorite.interactors
+
+import kotlinx.coroutines.flow.Flow
+import ru.practicum.android.diploma.domain.favorite.repository.GetFavoriteVacanciesListRepository
+import ru.practicum.android.diploma.domain.models.Vacancy
+
+class GetFavoriteVacanciesInteractorListImpl(
+    private val getFavoriteVacanciesListRepository: GetFavoriteVacanciesListRepository
+) : GetFavoriteVacanciesListInteractor {
+    override fun getFavVacanciesList(): Flow<List<Vacancy>> {
+        return getFavoriteVacanciesListRepository.getFavVacanciesList()
+    }
+}
