+++ app/src/main/java/ru/practicum/android/diploma/domain/favorite/repository/GetFavoriteVacanciesListRepository.kt
@@ -1,0 +1,8 @@
+package ru.practicum.android.diploma.domain.favorite.repository
+
+import kotlinx.coroutines.flow.Flow
+import ru.practicum.android.diploma.domain.models.Vacancy
+
+interface GetFavoriteVacanciesListRepository {
+    fun getFavVacanciesList(): Flow<List<Vacancy>>
+}
